Move all the python bindings into a separate library
