+++ src/python_bindings/pymatching.cc
@@ -13,9 +13,9 @@
 // limitations under the License.
 
 #include "pybind11/pybind11.h"
-#include "pymatching/rand/rand_gen.pybind.h"
+#include "python_bindings/rand_gen.h"
+#include "python_bindings/user_graph.h"
 #include "blossom/driver/namespaced_main.h"
-#include "blossom/driver/user_graph.pybind.h"
 
 namespace py = pybind11;
 using namespace py::literals;
